Moved groups_shouldBeVisibleInHierarchyView to FormHierarchyTest
